pyproject.toml
[tool.poetry]
name = "awsssomanager"
version = "0.1.0"
description = "AWS SSO Manager"
authors = ["Brett Fitzpatrick"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
coloredlogs = "^15.0.1"
boto3 = "^1.28.62"
pyyaml = "^6.0.1"
yamale = "^4.0.4"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocstrings = "^0.23.0"
mkdocs-material = "^9.4.4"
mkdocs-minify-plugin = "^0.7.1"
mkdocs-table-reader-plugin = "^2.0.1"

[tool.poetry.scripts]
aws-sso-manager = "awsssomanager.cli.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
